Changes how callers interact with framework.
Introduced a subclass of AFJSONRequestOperation called OHResourceRequestOperation. This is now used to encapsulate the asynchronous networking operations. Effect is to simplify the code and remove the need for our own NSOperationQueue.  OHClient now communicates with caller via KVO.  Test demonstrating use of KVO is included.

Started modification to how OHResource will be used to traverse embedded and linked resources.