Tests all pass. **Timing Issue Remains**
All of the traversal tests pass. This commit marks a working version,
but with timing issues. It should not be relied upon until the timing
issue is addressed.

The approach taken for traversal right now relies on a
single NSOperationQueue and managing the dependent NSOperations may not
be entirely correct. A modified approach where the OHLinkTraverser
derives from NSOperation and maintains an internal NSOperationQueue
would allow for truly dependent child operations.